requirements.txt
gradio==4.44.0
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76
scikit-learn==1.3.0
scikit-image==0.21.0